+++ src/Frontend/Desktop/Services/HerbService.cs
@@ -85,7 +85,7 @@ namespace LYBT.WPF.Client.Services
                     return new PagedResult
                     {
                         Items = herbInfos,
-                        TotalCount = response.Content.TotalCount,
+                        TotalCount = (int)response.Content.TotalCount,
                         CurrentPage = response.Content.CurrentPage,
                         PageSize = response.Content.PageSize
                     };
